# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
...
  dto.NurseResponse:
    properties:
      created_at:
        type: string
      department_id:
        type: string
      license_number:
        type: string
      nurse_id:
        type: string
      shift:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  dto.NurseSignupRequest:
    properties:
      department_id:
        type: string
      license_number:
        type: string
      shift:
        type: string
      user_id:
        type: string
    required:
    - department_id
    - license_number
    - shift
    - user_id
    type: object
  dto.PatientResponse:
    properties:
      blood_group:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      emergency_contact_name:
        type: string
      emergency_contact_phone:
        type: string
      gender:
        type: string
      medical_history:
        type: string
      patient_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  dto.PatientSignUp:
    properties:
      blood_group:
        type: string
      date_of_birth:
        type: string
      emergency_contact_name:
        type: string
      emergency_contact_phone:
        type: string
      gender:
        type: string
      medical_history:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@hms.example.com
    name: HMS Support
    url: http://example.com/support
  description: A comprehensive Hospital Management System with user authentication,
    doctor management, and patient care features
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: Hospital Management System API
  version: 1.0.0
paths:
...
  /admin/nurses:
    post:
      consumes:
      - application/json
      description: Create a new nurse. Requires valid JWT token with ADMIN role
      parameters:
      - description: Nurse creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.NurseSignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Nurse created successfully
          schema:
            $ref: '#/definitions/dto.NurseResponse'
        "400":
          description: Validation error - invalid input or invalid role
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict - nurse already exists for this user
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new nurse
      tags:
      - Nurse Management
...
  /patients/patientprofile:
    post:
      consumes:
      - application/json
      description: Create patient profile. Requires valid JWT token with PATIENT role
      parameters:
      - description: Patient Profile details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PatientSignUp'
      produces:
      - application/json
      responses:
        "201":
          description: Patient Profile created successfully
          schema:
            $ref: '#/definitions/dto.PatientResponse'
        "400":
          description: Validation error - invalid input or invalid role
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - patient role required
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict - patient already exists for this user
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create patient profile
      tags:
      - Patient Management
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOiJIUzI1NiIs..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"